kosan_backend/requirements.txt
asgiref==3.11.0
Django==5.2.8
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.7
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.15.0
